refactor: extract get_files_containing_symbol cluster to QueryEngine (Phase 3)
Move get_files_containing_symbol and 5 helpers from CppAnalyzer to QueryEngine:
- _find_class_definition_files
- _find_function_definition_files
- _find_symbol_definition_files
- _find_symbol_caller_files
- _find_class_reference_files
- get_files_containing_symbol (async)

CppAnalyzer keeps 1-line delegation for backward compatibility.
All state access in QueryEngine goes through self.analyzer back-reference.

All 1435 tests pass.
