+++ mcp_server/cpp_analyzer.py
@@ -2078,188 +2078,13 @@ class CppAnalyzer:
         """Search for symbols within a specific file or files matching a glob pattern (delegates to query_engine)."""
         return self.query_engine.find_in_file(file_path, pattern)
 
-    def _find_class_definition_files(
-        self,
-        symbol_name: str,
-        symbol_kind: Optional[str],
-        simple_name: str,
-        project_only: bool,
-        files: Set[str],
-    ) -> Optional[str]:
-        """Find files where the class is defined and return its kind."""
-        if symbol_kind in (None, "class"):
-            for info in self.class_index.get(simple_name, []):
-                if SearchEngine.matches_qualified_pattern(
-                    info.qualified_name or info.name, symbol_name
-                ):
-                    if not project_only or info.is_project:
-                        files.add(info.file)
-                        if info.header_file:
-                            files.add(info.header_file)
-                        return info.kind
-        return None
-
-    def _find_function_definition_files(
-        self,
-        symbol_name: str,
-        symbol_kind: Optional[str],
-        simple_name: str,
-        project_only: bool,
-        files: Set[str],
-    ) -> Optional[str]:
-        """Find files where the function/method is defined and return its kind."""
-        kind = None
-        if symbol_kind in (None, "function", "method"):
-            for info in self.function_index.get(simple_name, []):
-                if SearchEngine.matches_qualified_pattern(
-                    info.qualified_name or info.name, symbol_name
-                ):
-                    if not project_only or info.is_project:
-                        files.add(info.file)
-                        if info.header_file:
-                            files.add(info.header_file)
-                        if not kind:
-                            kind = info.kind
-        return kind
-
-    def _find_symbol_definition_files(
-        self,
-        symbol_name: str,
-        symbol_kind: Optional[str],
-        simple_name: str,
-        project_only: bool,
-        files: Set[str],
-    ) -> Optional[str]:
-        """Find files where the symbol is defined and return the first found kind."""
-        kind = self._find_class_definition_files(
-            symbol_name, symbol_kind, simple_name, project_only, files
-        )
-
-        func_kind = self._find_function_definition_files(
-            symbol_name, symbol_kind, simple_name, project_only, files
-        )
-
-        return kind or func_kind
-
-    def _find_symbol_caller_files(
-        self,
-        symbol_name: str,
-        symbol_kind: Optional[str],
-        simple_name: str,
-        project_only: bool,
-        kind: Optional[str],
-        files: Set[str],
-    ) -> int:
-        """Find files that call the symbol and return the reference count."""
-        total_refs = 0
-        if kind in ("function", "method") or (
-            not kind and symbol_kind in (None, "function", "method")
-        ):
-
-            def _name_matches(info) -> bool:
-                return SearchEngine.matches_qualified_pattern(
-                    info.qualified_name or info.name, symbol_name
-                )
-
-            # Get USRs for all functions with this name
-            target_usrs = set()
-            for info in self.function_index.get(simple_name, []):
-                if _name_matches(info) and info.usr:
-                    if not project_only or info.is_project:
-                        target_usrs.add(info.usr)
-
-            # Find all callers of these functions
-            for usr in target_usrs:
-                callers = self.call_graph_analyzer.find_incoming_calls(usr)
-                for caller_usr in callers:
-                    if caller_usr in self.usr_index:
-                        caller_info = self.usr_index[caller_usr]
-                        if not project_only or caller_info.is_project:
-                            files.add(caller_info.file)
-                            total_refs += 1
-        return total_refs
-
-    def _find_class_reference_files(
-        self,
-        symbol_name: str,
-        symbol_kind: Optional[str],
-        project_only: bool,
-        kind: Optional[str],
-        files: Set[str],
-    ) -> None:
-        """Find files that reference a class and add them to the set."""
-        if kind in ("class", "struct") or (not kind and symbol_kind in (None, "class")):
-            # Check file index for files that might reference the class
-            for file_path, symbols in self.file_index.items():
-                if not project_only or self.compilation_env._is_project_file(file_path):
-                    # If file has the class definition or any methods of the class
-                    for symbol in symbols:
-                        sym_qname = symbol.qualified_name or symbol.name
-                        parent_qname = symbol.parent_class or ""
-                        if SearchEngine.matches_qualified_pattern(
-                            sym_qname, symbol_name
-                        ) or SearchEngine.matches_qualified_pattern(parent_qname, symbol_name):
-                            files.add(file_path)
-                            break
-
     async def get_files_containing_symbol(
         self, symbol_name: str, symbol_kind: Optional[str] = None, project_only: bool = True
     ) -> Dict[str, Any]:
-        """
-        Get all files that contain references to or define a symbol.
-
-        Phase 1: LLM Integration - Enables targeted code search by narrowing down
-        which files to examine with filesystem or ripgrep MCP tools.
-
-        Args:
-            symbol_name: Name of the symbol (exact match, case-sensitive)
-            symbol_kind: Optional filter: "class", "function", or "method"
-            project_only: If True, exclude dependency/system files
-
-        Returns:
-            Dictionary with:
-                - symbol: The symbol name searched
-                - kind: The symbol kind if found (class, function, method)
-                - files: Sorted list of file paths containing the symbol
-                - total_references: Approximate count of references
-        """
-        # Note: Indexing should be complete before calling this method
-        files: Set[str] = set()
-        total_refs = 0
-        kind = None
-
-        # Support partially qualified names (e.g. "ClassName::method") by using the
-        # simple (unqualified) name for index keying and matches_qualified_pattern for
-        # the full match check.  This mirrors the approach used in find_incoming_calls/find_callees.
-        simple_name = symbol_name.split("::")[-1]
-
-        with self.index_lock:
-            # 1 & 2. Find where symbol is defined
-            kind = self._find_symbol_definition_files(
-                symbol_name, symbol_kind, simple_name, project_only, files
-            )
-
-            # 3. Find callers (for functions/methods)
-            total_refs = self._find_symbol_caller_files(
-                symbol_name, symbol_kind, simple_name, project_only, kind, files
-            )
-
-            # 4. For classes, find files that use the class
-            self._find_class_reference_files(symbol_name, symbol_kind, project_only, kind, files)
-
-        # Filter and sort files
-        file_list = sorted(list(files))
-
-        # If no references found via call graph, use file count as estimate
-        if total_refs == 0:
-            total_refs = len(file_list)
-
-        return {
-            "symbol": symbol_name,
-            "kind": kind,
-            "files": file_list,
-            "total_references": total_refs,
-        }
+        """Get all files that contain references to or define a symbol (delegates to query_engine)."""
+        return await self.query_engine.get_files_containing_symbol(
+            symbol_name, symbol_kind, project_only
+        )
 
     def get_parse_errors(
         self, limit: Optional[int] = None, file_path_filter: Optional[str] = None
